requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asyncer==0.0.7
attrs==23.2.0
autocommand==2.2.2
Babel==2.14.0
backoff==2.2.1
bcc==0.30.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.8.2
Brlapi==0.8.5
build==1.2.2.post1
cachetools==5.5.2
certifi==2024.6.2
chainlit==2.2.1
chardet==5.2.0
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
cloud-init==24.3.1
colorama==0.4.6
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.8
cryptography==42.0.5
cupshelpers==1.0
dataclasses-json==0.6.7
dbus-python==1.3.2
defer==1.0.6
Deprecated==1.2.18
distro==1.9.0
distro-info==1.9
dlib==19.24.5
durationpy==0.9
fastapi==0.115.11
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
greenlet==3.1.1
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
gyp-next==0.16.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
humanfriendly==10.0
idna==3.6
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflect==7.3.1
jaraco.context==6.0.0
jaraco.functools==4.0.2
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.0
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
kubernetes==32.0.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.43
langchain-google-genai==2.0.10
langchain-groq==0.2.5
langchain-text-splitters==0.3.6
langsmith==0.3.13
language-selector==0.1
launchpadlib==2.0.0
Lazify==0.4.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
literalai==0.1.103
louis==3.30.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netaddr==0.10.1
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.20.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.1
pexpect==4.9.0
pillow==10.4.0
posthog==3.19.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.26.1
pycups==2.0.4
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.18.0
PyGObject==3.48.2
PyJWT==2.10.1
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
pyserial==3.5
python-apt==2.9.0+ubuntu1
python-dateutil==2.9.0
python-debian==0.1.49+ubuntu3
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
pyxdg==0.28
PyYAML==6.0.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
screen-resolution-extra==0.0.0
sentence-transformers==3.4.1
setuptools==75.8.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.41.3
sympy==1.13.1
syncer==2.0.3
systemd-python==235
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typeguard==4.3.0
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uptrace==1.29.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.34.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==0.20.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.44.0
wrapt==1.17.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
